+++ libraries/matchers.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 if defined?(ChefSpec)
   def create_osquery_config(conf)
     ChefSpec::Matchers::ResourceMatcher.new(:osquery_conf, :create, conf)