+++ resources/install.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 actions :install_ubuntu, :install_centos, :install_os_x, :remove_centos, :remove_ubuntu, :remove_os_x,
         :install_amazon, :remove_amazon
 default_action :install_linux