+++ spec/unit/resources/osquery_syslog_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe 'osquery_spec::osquery_syslog' do