+++ test/integration/default/inspec/osquery_spec.rb
@@ -1,17 +1,4 @@
-expected_options = {
-    'syslog_pipe_path' => "/var/osquery/syslog_pipe",
-    'config_plugin' => 'filesystem',
-    'logger_plugin' => 'filesystem',
-    'schedule_splay_percent' =>  10,
-    'events_expiry' => 3600,
-    'verbose' => false,
-    'worker_threads' => 2,
-    'logger_path' => "/var/log/osquery"
-}
-
-expected_schedule = {
-    'info' => {'query' => 'SELECT * FROM osquery_info;', 'interval' => 86400}
-}
+# frozen_string_literal: true
 
 case os[:family]
 when 'debian', 'redhat', 'amazon'
@@ -25,6 +12,29 @@ when 'debian', 'redhat', 'amazon'
     it { should exist }
     it { should be_file }
     its('mode') { should cmp '0440' }
+  end
+
+  describe json('/etc/osquery/osquery.conf') do
+    expected_options = {
+      'syslog_pipe_path' => '/var/osquery/syslog_pipe',
+      'config_plugin' => 'filesystem',
+      'logger_plugin' => 'filesystem',
+      'schedule_splay_percent' => 10,
+      'events_expiry' => 3600,
+      'verbose' => false,
+      'worker_threads' => 2,
+      'logger_path' => '/var/log/osquery'
+    }
+
+    expected_schedule = {
+      'info' => { 'query' => 'SELECT * FROM osquery_info;', 'interval' => 86_400 }
+    }
+
+    its(['options']) { should eq expected_options }
+    its(['schedule']) { should eq expected_schedule }
+    its(['decorators']) { should eq nil }
+    its(['file_paths']) { should eq nil }
+    its(['exclude_paths']) { should eq nil }
   end
 
   describe file('/usr/share/osquery/packs') do
@@ -44,15 +54,6 @@ when 'debian', 'redhat', 'amazon'
     its('version') { should eq '4.0.2-1.linux' }
   end
 
-  describe json("/etc/osquery/osquery.conf") do
-    its(['options']) {should eq expected_options}
-    its(['schedule']) {should eq expected_schedule}
-    its(['decorators']) {should eq nil}
-    its(['file_paths']) {should eq nil}
-    its(['exclude_paths']) {should eq nil}
-  end
-
-
 when 'darwin'
   describe file('/var/osquery/osquery.conf') do
     it { should exist }
@@ -64,5 +65,28 @@ when 'darwin'
     it { should be_installed }
     it { should be_enabled }
     it { should be_running }
+  end
+
+  describe json('/var/osquery/osquery.conf') do
+    expected_options = {
+      'syslog_pipe_path' => '/var/osquery/syslog_pipe',
+      'config_plugin' => 'filesystem',
+      'logger_plugin' => 'filesystem',
+      'schedule_splay_percent' => 10,
+      'events_expiry' => 3600,
+      'verbose' => false,
+      'worker_threads' => 2,
+      'logger_path' => '/var/log/osquery'
+    }
+
+    expected_schedule = {
+      'info' => { 'query' => 'SELECT * FROM osquery_info;', 'interval' => 86_400 }
+    }
+
+    its(['options']) { should eq expected_options }
+    its(['schedule']) { should eq expected_schedule }
+    its(['decorators']) { should eq nil }
+    its(['file_paths']) { should eq nil }
+    its(['exclude_paths']) { should eq nil }
   end
 end