+++ libraries/helpers.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'mixlib/shellout'
 
 # helper methods for common case statements
@@ -10,20 +12,20 @@ module Osquery
     case node['platform']
     when 'mac_os_x'
       'com.facebook.osqueryd'
-    when 'centos', 'ubuntu', 'redhat'
+    when 'centos', 'ubuntu', 'redhat', 'amazon'
       'osqueryd'
     end
   end
 
   def osquery_s3
-    'https://osquery-packages.s3.amazonaws.com'
+    'https://pkg.osquery.io'
   end
 
   def osquery_config_path
     case node['platform']
     when 'mac_os_x'
       '/var/osquery/osquery.conf'
-    when 'centos', 'ubuntu', 'redhat'
+    when 'centos', 'ubuntu', 'redhat', 'amazon'
       '/etc/osquery/osquery.conf'
     end
   end
@@ -32,7 +34,7 @@ module Osquery
     case node['platform']
     when 'mac_os_x'
       '/var/osquery/packs'
-    when 'centos', 'ubuntu', 'redhat'
+    when 'centos', 'ubuntu', 'redhat', 'amazon'
       '/usr/share/osquery/packs'
     end
   end
@@ -41,7 +43,7 @@ module Osquery
     case node['platform']
     when 'mac_os_x'
       'wheel'
-    when 'centos', 'ubuntu', 'redhat'
+    when 'centos', 'ubuntu', 'redhat', 'amazon'
       'root'
     end
   end
@@ -51,6 +53,6 @@ module Osquery
   end
 end
 
-Chef::Recipe.send(:include, Osquery)
-Chef::Resource.send(:include, Osquery)
-Chef::Provider.send(:include, Osquery)
+Chef::Recipe.include Osquery
+Chef::Resource.include Osquery
+Chef::Provider.include Osquery