+++ spec/fixtures/osquery_spec/recipes/osquery_conf.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 #
 # Cookbook Name:: osquery_spec
 # Recipe:: default
@@ -15,7 +17,7 @@ node.default['osquery']['decorators']['always'] = [
   "SELECT 'production' as envIdentifier"
 ]
 node.override['osquery']['pack_source'] = 'osquery_spec'
-node.override['osquery']['packs'] = %w(osquery_spec_test)
+node.override['osquery']['packs'] = %w[osquery_spec_test]
 
 osquery_conf osquery_config_path do
   action      :create