+++ spec/fixtures/osquery_spec/recipes/osquery_decorators.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 #
 # Cookbook Name:: osquery_spec
 # Recipe:: default
@@ -9,6 +11,6 @@ node.default['osquery']['decorators']['always'] = [
   "SELECT 'production' as envIdentifier"
 ]
 node.override['osquery']['pack_source'] = 'osquery_spec'
-node.override['osquery']['packs'] = %w(osquery_spec_test)
+node.override['osquery']['packs'] = %w[osquery_spec_test]
 
 include_recipe 'osquery::default'