+++ spec/unit/recipes/default_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe 'osquery::default' do
@@ -5,7 +7,7 @@ describe 'osquery::default' do
     include_context 'converged recipe'
 
     let(:platform) do
-      { platform: 'mac_os_x', version: '10.10' }
+      { platform: 'mac_os_x', version: '10.12' }
     end
 
     it 'includes mac os x installation recipe' do
@@ -37,7 +39,7 @@ describe 'osquery::default' do
     include_context 'converged recipe'
 
     let(:platform) do
-      { platform: 'centos', version: '7.0' }
+      { platform: 'centos', version: '7.7.1908' }
     end
 
     it 'includes centos installation recipe' do
@@ -53,11 +55,27 @@ describe 'osquery::default' do
     include_context 'converged recipe'
 
     let(:platform) do
-      { platform: 'redhat', version: '7.0' }
+      { platform: 'redhat', version: '7.6' }
     end
 
     it 'includes centos installation recipe' do
       expect(chef_run).to include_recipe('osquery::centos')
+    end
+
+    it 'converges without error' do
+      expect { chef_run }.not_to raise_error
+    end
+  end
+
+  context 'when amazon' do
+    include_context 'converged recipe'
+
+    let(:platform) do
+      { platform: 'amazon', version: '2018.03' }
+    end
+
+    it 'includes amazon installation recipe' do
+      expect(chef_run).to include_recipe('osquery::amazon')
     end
 
     it 'converges without error' do